remove annotatedType parameter and supplier, optional property
