+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/OptionalContainerPropertyGenerator.java
@@ -20,8 +20,10 @@ package com.navercorp.fixturemonkey.api.generator;
 
 import static com.navercorp.fixturemonkey.api.type.Types.generateAnnotatedTypeWithoutAnnotation;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Type;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -29,10 +31,11 @@ import java.util.OptionalDouble;
 import java.util.OptionalInt;
 import java.util.OptionalLong;
 
+import javax.annotation.Nullable;
+
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
-import com.navercorp.fixturemonkey.api.property.OptionalProperty;
 import com.navercorp.fixturemonkey.api.property.Property;
 import com.navercorp.fixturemonkey.api.property.SingleElementProperty;
 import com.navercorp.fixturemonkey.api.type.Types;
@@ -53,8 +56,41 @@ public final class OptionalContainerPropertyGenerator implements ContainerProper
 		AnnotatedType valueAnnotatedType = getOptionalValueAnnotatedType(property);
 		Type valueType = valueAnnotatedType.getType();
 
-		Property childProperty = new OptionalProperty(valueType, valueAnnotatedType);
-		SingleElementProperty singleElementProperty = new SingleElementProperty(childProperty, valueAnnotatedType);
+		Property childProperty = new Property() {
+			@Override
+			public Type getType() {
+				return valueType;
+			}
+
+			@Override
+			public AnnotatedType getAnnotatedType() {
+				return valueAnnotatedType;
+			}
+
+			@Nullable
+			@Override
+			public String getName() {
+				return null;
+			}
+
+			@Override
+			public List<Annotation> getAnnotations() {
+				return Arrays.asList(valueAnnotatedType.getAnnotations());
+			}
+
+			@Nullable
+			@Override
+			public Object getValue(Object instance) {
+				Class<?> actualType = Types.getActualType(instance.getClass());
+				if (isOptional(actualType)) {
+					return getOptionalValue(instance);
+				}
+
+				throw new IllegalArgumentException("given value has no match");
+			}
+		};
+
+		SingleElementProperty singleElementProperty = new SingleElementProperty(childProperty);
 
 		return new ContainerProperty(
 			Collections.singletonList(singleElementProperty),
@@ -92,5 +128,33 @@ public final class OptionalContainerPropertyGenerator implements ContainerProper
 		}
 
 		return genericsTypes.get(0);
+	}
+
+	private boolean isOptional(Class<?> type) {
+		return Optional.class.isAssignableFrom(type)
+			|| OptionalInt.class.isAssignableFrom(type)
+			|| OptionalLong.class.isAssignableFrom(type)
+			|| OptionalDouble.class.isAssignableFrom(type);
+	}
+
+	private Object getOptionalValue(Object obj) {
+		Class<?> actualType = Types.getActualType(obj.getClass());
+		if (Optional.class.isAssignableFrom(actualType)) {
+			return ((Optional<?>)obj).orElse(null);
+		}
+
+		if (OptionalInt.class.isAssignableFrom(actualType)) {
+			return ((OptionalInt)obj).orElse(0);
+		}
+
+		if (OptionalLong.class.isAssignableFrom(actualType)) {
+			return ((OptionalLong)obj).orElse(0L);
+		}
+
+		if (OptionalDouble.class.isAssignableFrom(actualType)) {
+			return ((OptionalDouble)obj).orElse(Double.NaN);
+		}
+
+		throw new IllegalArgumentException("given value is not optional, actual type : " + actualType);
 	}
 }
