+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/generator/SupplierContainerPropertyGenerator.java
@@ -18,17 +18,20 @@
 
 package com.navercorp.fixturemonkey.api.generator;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Type;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Supplier;
+
+import javax.annotation.Nullable;
 
 import org.apiguardian.api.API;
 
 import com.navercorp.fixturemonkey.api.property.Property;
 import com.navercorp.fixturemonkey.api.property.SingleElementProperty;
-import com.navercorp.fixturemonkey.api.property.SupplierProperty;
 import com.navercorp.fixturemonkey.api.type.Types;
 
 @API(since = "1.0.16", status = API.Status.EXPERIMENTAL)
@@ -43,8 +46,41 @@ public final class SupplierContainerPropertyGenerator implements ContainerProper
 		AnnotatedType valueAnnotatedType = getSupplierValueAnnotatedType(property);
 		Type valueType = valueAnnotatedType.getType();
 
-		Property childProperty = new SupplierProperty(valueType, valueAnnotatedType);
-		SingleElementProperty singleElementProperty = new SingleElementProperty(childProperty, valueAnnotatedType);
+		Property childProperty = new Property() {
+			@Override
+			public Type getType() {
+				return valueType;
+			}
+
+			@Override
+			public AnnotatedType getAnnotatedType() {
+				return valueAnnotatedType;
+			}
+
+			@Nullable
+			@Override
+			public String getName() {
+				return null;
+			}
+
+			@Override
+			public List<Annotation> getAnnotations() {
+				return Arrays.asList(valueAnnotatedType.getAnnotations());
+			}
+
+			@Override
+			public Object getValue(Object instance) {
+				Class<?> actualType = Types.getActualType(instance.getClass());
+
+				if (Supplier.class.isAssignableFrom(actualType)) {
+					return instance;
+				}
+
+				throw new IllegalArgumentException("given value has no match");
+			}
+		};
+
+		SingleElementProperty singleElementProperty = new SingleElementProperty(childProperty);
 
 		return new ContainerProperty(
 			Collections.singletonList(singleElementProperty),
