--- fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/property/OptionalProperty.java
@@ -1,89 +1,0 @@
-package com.navercorp.fixturemonkey.api.property;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.AnnotatedType;
-import java.lang.reflect.Type;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.OptionalDouble;
-import java.util.OptionalInt;
-import java.util.OptionalLong;
-
-import javax.annotation.Nullable;
-
-import org.apiguardian.api.API;
-
-import com.navercorp.fixturemonkey.api.type.Types;
-
-@API(since = "1.0.16", status = API.Status.EXPERIMENTAL)
-public final class OptionalProperty implements Property {
-
-	private final Type type;
-	private final AnnotatedType annotatedType;
-
-	public OptionalProperty(Type type, AnnotatedType annotatedType) {
-		this.type = type;
-		this.annotatedType = annotatedType;
-	}
-
-	@Override
-	public Type getType() {
-		return type;
-	}
-
-	@Override
-	public AnnotatedType getAnnotatedType() {
-		return annotatedType;
-	}
-
-	@Nullable
-	@Override
-	public String getName() {
-		return null;
-	}
-
-	@Override
-	public List<Annotation> getAnnotations() {
-		return Arrays.asList(annotatedType.getAnnotations());
-	}
-
-	@Nullable
-	@Override
-	public Object getValue(Object instance) {
-		Class<?> actualType = Types.getActualType(instance.getClass());
-		if (isOptional(actualType)) {
-			return getOptionalValue(instance);
-		}
-
-		throw new IllegalArgumentException("given value has no match");
-	}
-
-	private boolean isOptional(Class<?> type) {
-		return Optional.class.isAssignableFrom(type)
-			|| OptionalInt.class.isAssignableFrom(type)
-			|| OptionalLong.class.isAssignableFrom(type)
-			|| OptionalDouble.class.isAssignableFrom(type);
-	}
-
-	private Object getOptionalValue(Object obj) {
-		Class<?> actualType = Types.getActualType(obj.getClass());
-		if (Optional.class.isAssignableFrom(actualType)) {
-			return ((Optional<?>)obj).orElse(null);
-		}
-
-		if (OptionalInt.class.isAssignableFrom(actualType)) {
-			return ((OptionalInt)obj).orElse(0);
-		}
-
-		if (OptionalLong.class.isAssignableFrom(actualType)) {
-			return ((OptionalLong)obj).orElse(0L);
-		}
-
-		if (OptionalDouble.class.isAssignableFrom(actualType)) {
-			return ((OptionalDouble)obj).orElse(Double.NaN);
-		}
-
-		throw new IllegalArgumentException("given value is not optional, actual type : " + actualType);
-	}
-}
